Address feedback on tests and improve them to cover more scenarios.

+++ src/libraries/Microsoft.Extensions.FileProviders.Physical/tests/PhysicalFileProviderTests.netcoreapp.cs
@@ -15,29 +15,6 @@ namespace Microsoft.Extensions.FileProviders
         [Theory]
         [InlineData(false)]
         [InlineData(true)]
-        public void UsePollingFileWatcher_UseActivePolling_HasChanged(bool useWildcard)
-        {
-            // Arrange
-            using var root = new DisposableFileSystem();
-            string fileName = Path.GetRandomFileName();
-            string filePath = Path.Combine(root.RootPath, fileName);
-            File.WriteAllText(filePath, "v1.1");
-
-            using var provider = new PhysicalFileProvider(root.RootPath) { UsePollingFileWatcher = true, UseActivePolling = true };
-            IChangeToken token = provider.Watch(useWildcard ? "*" : fileName);
-            Assert.False(token.HasChanged);
-
-            // Act
-            File.WriteAllText(filePath, "v1.2");
-            Thread.Sleep(GetTokenPollingInterval(token));
-
-            // Assert
-            Assert.True(token.HasChanged);
-        }
-
-        [Theory]
-        [InlineData(false)]
-        [InlineData(true)]
         public void UsePollingFileWatcher_UseActivePolling_HasChanged_SymbolicLink(bool useWildcard)
         {
             // Arrange
@@ -55,6 +32,7 @@ namespace Microsoft.Extensions.FileProviders
             Assert.False(token.HasChanged);
 
             // Act
+            Thread.Sleep(100); // Wait a bit before writing again, see https://github.com/dotnet/runtime/issues/55951.
             File.WriteAllText(filePath, "v1.2");
             Thread.Sleep(GetTokenPollingInterval(token));
 
@@ -82,17 +60,24 @@ namespace Microsoft.Extensions.FileProviders
         }
 
         [Theory]
-        [InlineData(false)]
-        [InlineData(true)]
-        public void UsePollingFileWatcher_UseActivePolling_HasChanged_SymbolicLink_TargetChanged(bool useWildcard)
+        [InlineData(false, false)]
+        [InlineData(false, true)]
+        [InlineData(true, false)]
+        [InlineData(true, true)]
+        public void UsePollingFileWatcher_UseActivePolling_HasChanged_SymbolicLink_TargetChanged(bool useWildcard, bool fromTargetNonExistent)
         {
             // Arrange
             using var rootOfFile = new DisposableFileSystem();
-            string file1Path = Path.Combine(rootOfFile.RootPath, Path.GetRandomFileName());
-            File.WriteAllText(file1Path, "v1.1");
-
+            // Create file 2 first as we want to verify that the change is reported regardless of the timestamp being older.
             string file2Path = Path.Combine(rootOfFile.RootPath, Path.GetRandomFileName());
             File.WriteAllText(file2Path, "v2.1");
+
+            string file1Path = Path.Combine(rootOfFile.RootPath, Path.GetRandomFileName());
+            if (!fromTargetNonExistent)
+            {
+                Thread.Sleep(100); // Wait a bit before writing again, see https://github.com/dotnet/runtime/issues/55951.
+                File.WriteAllText(file1Path, "v1.1");
+            }
 
             using var rootOfLink = new DisposableFileSystem();
             string linkName = Path.GetRandomFileName();
@@ -104,43 +89,42 @@ namespace Microsoft.Extensions.FileProviders
             IChangeToken token = provider.Watch(filter);
             Assert.False(token.HasChanged);
 
-            // Act 1 - Change file 1's content.
-            File.WriteAllText(file1Path, "v1.2");
-            Thread.Sleep(GetTokenPollingInterval(token));
-
-            // Assert 1
-            Assert.True(token.HasChanged);
-
-            // Act 2 - Change link target to file 2.
-            token = provider.Watch(filter); // Once HasChanged is true, the value will always be true. Get a new change token.
-            Assert.False(token.HasChanged);
+            // Act - Change link target to file 2.
             File.Delete(linkPath);
             File.CreateSymbolicLink(linkPath, file2Path);
             Thread.Sleep(GetTokenPollingInterval(token));
 
-            // Assert 2
+            // Assert
             Assert.True(token.HasChanged); // It should report the change regardless of the timestamp being older.
-
-            // Act 3 - Change file 2's content.
-            token = provider.Watch(filter);
-            Assert.False(token.HasChanged);
-            File.WriteAllText(file2Path, "v2.2");
-            Thread.Sleep(GetTokenPollingInterval(token));
-
-            // Assert 3
-            Assert.True(token.HasChanged);
         }
 
-        private int GetTokenPollingInterval(IChangeToken changeToken)
+        [Theory]
+        [InlineData(false)]
+        [InlineData(true)]
+        public void UsePollingFileWatcher_UseActivePolling_HasChanged_SymbolicLink_TargetDeleted(bool useWildcard)
         {
-            TimeSpan pollingInterval = (changeToken as CompositeChangeToken).ChangeTokens[1] switch
-            {
-                PollingWildCardChangeToken wildcardChangeToken => wildcardChangeToken.PollingInterval,
-                PollingFileChangeToken => PollingFileChangeToken.PollingInterval,
-                _ => throw new InvalidOperationException()
-            };
+            // Arrange
+            using var rootOfFile = new DisposableFileSystem();
 
-            return (int)pollingInterval.TotalMilliseconds;
+            string filePath = Path.Combine(rootOfFile.RootPath, Path.GetRandomFileName());
+            File.WriteAllText(filePath, "v1.1");
+
+            using var rootOfLink = new DisposableFileSystem();
+            string linkName = Path.GetRandomFileName();
+            string linkPath = Path.Combine(rootOfLink.RootPath, linkName);
+            File.CreateSymbolicLink(linkPath, filePath);
+
+            string filter = useWildcard ? "*" : linkName;
+            using var provider = new PhysicalFileProvider(rootOfLink.RootPath) { UsePollingFileWatcher = true, UseActivePolling = true };
+            IChangeToken token = provider.Watch(filter);
+            Assert.False(token.HasChanged);
+
+            // Act
+            File.Delete(linkPath);
+            Thread.Sleep(GetTokenPollingInterval(token));
+
+            // Assert
+            Assert.True(token.HasChanged);
         }
     }
 }
